// src/app/features/home/search/models/patient.ts
import { PatientDiseases } from "../../general-data/models/patient-model";

export interface Patient {

  pageSize?: number;
  pageIndex?: number;
  sortColumn?: string;
  sortOrder?: string;
  searchText?: string;
  id?: number;
  fullName?: string;
  incidentGovernmentId?: number;
  incidentGovernmentName?: string;
  incidentHealthAdministrationId?: number;
  incidentHealthAdministrationName?: string;
  incidentSourceId?: number;
  incidentDepartmentId?: number;
  nationalityId?: number;
  patientJobCategoryId?: number;
  patientJobId?: number;
  homeGovernmentId?: number;
  homeHealthAdministrationId?: number;
  homeCityId?: number;
  homeHealthOfficeId?: number;
  homePrincipalityId?: number;
  finalResultId?: number;
  transferGovernmentId?: number;
  transferIncidentSourceId?: number;
  labPlaceId?: number;
  caseResultCategoryId?: number;
  diseaseSeverityId?: number;
  labPrimaryDiagonistics?: number;
  caseDiscoveryDate?: Date;
  firstName?: string;
  secondName?: string;
  thirdName?: string;
  familyName?: string;
  nationalId?: string;
  passportNo?: string;
  phoneNo1?: string;
  phoneNo2?: string;
  genderId?: number;
  birthDate?: string;
  age?: number;
  ageTypeId?: number;
  maritalStatusId?: number;
  patientJobName?: string;
  educationPhaseId?: number;
  schoolCategoryId?: number;
  workAddress?: string;
  livingAddress?: string;
  patientHospitalNo?: string;
  doctorName?: string;
  hospitalEntryDate?: Date;
  hospitalLeaveDate?: Date;
  infectionDate?: Date;
  incidentDate?: Date;
  insertedByLab?: true;
  startDate?: Date;
  endDate?: Date;
  relativeTypeId?: number;
  expanded?: boolean;
  homeGovernmentName?: string;
  homeHealthAdministrationName?: string;
  homeHealthOfficeName?: string;
  homeCityName?: string;
  incidentSourceName?: string;
  nationality?: string;
  labChecksExistance?: string;
  totalCount?: number;
  filterType?: number;
  incidentBranchId?: number;
  incidentAreaId?: number;
}